+++ app/build.gradle
@@ -11,8 +11,8 @@ android {
         applicationId "com.cpa.cpa_word_problem"
         minSdkVersion 24
         targetSdkVersion 29
-        versionCode 10
-        versionName "1.2.2"
+        versionCode 11
+        versionName "1.3.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         kapt {
