--- app/release/output-metadata.json
@@ -1,20 +1,0 @@
-{
-  "version": 1,
-  "artifactType": {
-    "type": "APK",
-    "kind": "Directory"
-  },
-  "applicationId": "com.cpa.cpa_word_problem",
-  "variantName": "release",
-  "elements": [
-    {
-      "type": "SINGLE",
-      "filters": [],
-      "properties": [],
-      "versionCode": 10,
-      "versionName": "1.2.2",
-      "enabled": true,
-      "outputFile": "app-release.apk"
-    }
-  ]
-}
