+++ app/src/main/java/com/cpa/cpa_word_problem/adapters/MyFragmentStateAdapter.kt
@@ -3,9 +3,9 @@ package com.cpa.cpa_word_problem.adapters
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.cpa.cpa_word_problem.NoteFragment
-import com.cpa.cpa_word_problem.QuizFragment
-import com.cpa.cpa_word_problem.SettingFragment
+import com.cpa.cpa_word_problem.ui.NoteFragment
+import com.cpa.cpa_word_problem.ui.QuizFragment
+import com.cpa.cpa_word_problem.ui.SettingFragment
 
 class MyFragmentStateAdapter(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
