+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".ui.MainActivity">
 
         <androidx.viewpager2.widget.ViewPager2
             android:id="@+id/viewPager2"
