+++ app/src/main/res/layout/fragment_quiz.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QuizFragment">
+    tools:context=".ui.QuizFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/quizFragmentContainer"
