+++ app/src/main/res/layout/fragment_quiz_start.xml
@@ -5,7 +5,7 @@
     android:id="@+id/quizStartWholeLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".QuizStartFragment">
+    tools:context=".ui.QuizStartFragment">
 
     <TextView
         android:id="@+id/descriptionTextView"
