+++ app/src/main/res/layout/fragment_setting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:defaultFocusHighlightEnabled="true"
-    tools:context=".SettingFragment">
+    tools:context=".ui.SettingFragment">
 
     <TextView
         android:id="@+id/settingTitleView"
