+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">CPA 회계학 말문제</string>
+    <string name="app_name">CPA 말문제</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="START_YEAR">2016</string>
     <string name="END_YEAR">2020</string>
